Insurance_pdf_extractor-main/backend/requirements.txt
# Document Extraction Pipeline Requirements
# Successfully tested on Python 3.11.9

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==3.0.1

# Core Docling library (includes pandas, openpyxl, python-docx, pillow, tqdm, jsonschema)
docling==2.73.0

# Document Processing
pdf2image==1.16.3
PyPDF2==3.0.1
pypdf==6.7.0
pdfplumber==0.11.7

# OCR
pytesseract==0.3.10

# Utilities
requests>=2.32.2
python-dotenv==1.0.0

# Optional / Advanced Table Extraction
tabulate==0.9.0
camelot-py==0.11.0

rostaing-ocr


# Note: docling automatically installs compatible versions of:
# - pandas (>=2.2.2)
# - numpy
# - Pillow
# - openpyxl
# - python-docx
# - jsonschema
# - tqdm